Move asyncronous metrics state helper out of apimetric/metrictes
